requirements.txt
selenium==4.0.0-beta-4
python-dotenv
bpython
timer
